image_blur/requirements.txt
pillow==11.1.0
torch==2.6.0
numpy==2.2.3
opencv-python==4.11.0.86
transformers==4.49.0
fastapi[standard]==0.115.12
